Finalizing CRUDs
- Removed all Entity Services
- Removed Invoice Entity
- Removed Invoice_Item Entity

Decided Invoices will not be saved due to possible future complications

+ CustomError
+ CustomSuccess

These two classes will be used to return a success object or an error object to the renderer.
This makes it easy to change the way results and errors are handlers.

Finalized the CRUDs for
Customers, Units, Tax Types, Items, and Settings

Settings will be used for holding types of information like next invoice number. There might be more data that will be saved there in the future.

I've not yet decided where the creation of the CustomerIds will be handled. Most possibly frontend.